+++ src/schemathesis/specs/openapi/negative/utils.py
@@ -1,5 +1,7 @@
 from typing import List
 
+from hypothesis_jsonschema._canonicalise import canonicalish
+
 from .types import Schema
 
 
@@ -8,3 +10,7 @@ def get_type(schema: Schema) -> List[str]:
     if isinstance(type_, str):
         return [type_]
     return type_
+
+
+def can_negate(schema: Schema) -> bool:
+    return canonicalish(schema) != {}